+++ src/main/resources/db/migration/V6__posts_table_drop_followers_add_friends.sql
@@ -1,11 +1,11 @@
-DROP TABLE followers
+DROP TABLE followers;
 
-DROP TABLE followings
+DROP TABLE followings;
 
 CREATE TABLE friends (
   id bigserial PRIMARY KEY,
-  user_id INTEGER REFERENCES users (id) NOT NULL,
-  friend_id INTEGER REFERENCES users (id) NOT NULL,
+  user_id BIGINT REFERENCES users (id) NOT NULL,
+  friend_id BIGINT REFERENCES users (id) NOT NULL,
   accepted BOOLEAN NOT NULL,
   visibility TEXT NOT NULL,
   time TIMESTAMP NOT NULL
